Usually int is used for tile level variables, so use it here too.
svn path=/trunk/KDE/kdeedu/marble/; revision=1080568

+++ src/lib/global.h
@@ -232,7 +232,7 @@ const qreal RAD2INT = 21600.0 / M_PI;
 const qreal EARTH_RADIUS = 6378000.0;
 
 // Maximum level of base tiles
-const quint8 maxBaseTileLevel = 4;
+const int maxBaseTileLevel = 4;
 
 class MarbleGlobalPrivate;
 class MarbleLocale;
